Add some very basic rate limiting.
This is incredibly basic, and only serves to protect the authentication
endpoints, to hopefully prevent an attacker from just spamming login
attempts. The limit may still be far too generous for these endpoints,
however.

I didn't want to use a custom middleware, but every package I tried
(including slowapi) was seemingly unmaintained. Given this project is
largely intended for personal use, it's not a huge concern at this time.

Signed-off-by: Jason Lynch <jason@aexoden.com>